+++ components/remote-controls/setup/api/src/main/kotlin/com/flipperdevices/remotecontrols/api/DispatchSignalApi.kt
@@ -9,14 +9,17 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface DispatchSignalApi : InstanceKeeper.Instance {
     val state: StateFlow<State>
-    val isEmulated: StateFlow<Boolean>
 
     fun dismissBusyDialog()
 
     /**
      * Dispatch key from temporal file which contains only one key
      */
-    fun dispatch(config: EmulateConfig, identifier: IfrKeyIdentifier)
+    fun dispatch(
+        config: EmulateConfig,
+        identifier: IfrKeyIdentifier,
+        onDispatched: () -> Unit = {}
+    )
 
     fun reset()
 
@@ -26,7 +29,8 @@ interface DispatchSignalApi : InstanceKeeper.Instance {
     fun dispatch(
         identifier: IfrKeyIdentifier,
         remotes: List<InfraredRemote>,
-        ffPath: FlipperFilePath
+        ffPath: FlipperFilePath,
+        onDispatched: () -> Unit = {}
     )
 
     sealed interface State {