+++ components/remote-controls/setup/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/setup/composable/components/LoadedContent.kt
@@ -1,13 +1,8 @@
 package com.flipperdevices.remotecontrols.impl.setup.composable.components
 
 import android.content.res.Configuration
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -23,8 +18,6 @@ import com.flipperdevices.remotecontrols.setup.impl.R as SetupR
 @Composable
 fun LoadedContent(
     model: SetupComponent.Model.Loaded,
-    onPositiveClick: () -> Unit,
-    onNegativeClick: () -> Unit,
     onDispatchSignalClick: () -> Unit,
     modifier: Modifier = Modifier
 ) {
@@ -44,22 +37,6 @@ fun LoadedContent(
                     isSyncing = model.isSyncing,
                     isConnected = model.isConnected
                 )
-                AnimatedVisibility(
-                    visible = model.isEmulated,
-                    enter = slideInVertically(initialOffsetY = { it / 2 }),
-                    exit = slideOutVertically(targetOffsetY = { it / 2 }),
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .align(Alignment.BottomCenter)
-                        .navigationBarsPadding(),
-                ) {
-                    ConfirmContent(
-                        text = signalResponse.message.format(signalResponse.categoryName),
-                        onNegativeClick = onNegativeClick,
-                        onPositiveClick = onPositiveClick,
-                        modifier = Modifier.align(Alignment.BottomCenter)
-                    )
-                }
             }
 
             else -> {
@@ -82,12 +59,9 @@ private fun LoadedContentPreview() {
         LoadedContent(
             model = SetupComponent.Model.Loaded(
                 response = SignalResponseModel(),
-                isEmulated = true,
                 emulatedKeyIdentifier = null,
                 connectionState = InfraredConnectionApi.InfraredEmulateState.ALL_GOOD
             ),
-            onPositiveClick = {},
-            onNegativeClick = {},
             onDispatchSignalClick = {}
         )
     }